feat: 服务端 MySQL 权威存储 + 客户端 SQLite outbox 可靠重试 + 增量同步
- 服务端新增 user_message_sync 创建流表（迁移 20260813）：文本/系统消息写库同事务插双方同步行，图片在 ResourceServer 上传完成事务中插入
- 1051/1052 原地改名 ID_SYNC_MESSAGE_REQ/RSP，语义改为 bootstrap checkpoint + after_sync_seq 增量分页；删除 Redis offline_msg ZSET、GetPendingMessages、[Delivery] Offline* 配置
- 客户端新增 LocalChatStore/LocalChatDb/OutboxDispatcher/ChatSyncManager：每用户一个 SQLite（WAL+FULL），outbox 驱动 1017/1035/续传/1049 重试，先落库后上屏，插入去重保证推送与同步并发只展示一次
- TcpMgr 回归纯传输，删除 QSettings pending/replay/离线轮询全部旧实现
- 协议破坏性变更：message_id/thread_id/sync_seq 全链路 64 位、TCP JSON 十进制字符串；proto id 字段 int32→int64；两端须同批发布
- 测试：集成场景改写（同步补齐/多页/图片 1038 后可见/新增 sync-bootstrap、big-ids），新增 local_store_tests；bench 脚本与配置适配
- 构建：qt5-base 增加 sqlite3plugin feature，客户端链接 Qt5::Sql 并静态导入 qsqlite 驱动

+++ server/ResourceServer/include/FileWorker.h
@@ -16,7 +16,7 @@ struct FileTask {
 	FileTask(std::shared_ptr<CSession> session,  MSG_TYPES msg_type, int uid, std::string path, std::string name,
 		int seq, int total_size, int trans_size, int last,
 		std::string file_data,
-		std::function<void(const json&)> callback,int chat_msg_id=0,
+		std::function<void(const json&)> callback,long long chat_msg_id=0,
 		int sender = 0, int receiver = 0) :_session(session), _msg_type(msg_type),_uid(uid),
 		_seq(seq), _path(path), _name(name), _total_size(total_size),
 		_trans_size(trans_size), _last(last), _file_data(file_data), _callback(callback), _chat_msg_id(chat_msg_id),
@@ -34,10 +34,10 @@ struct FileTask {
 	int _last ;
 	std::string _file_data;
 	std::function<void(const json&)>  _callback;  //添加回调函数
-	int _chat_msg_id;
+	long long _chat_msg_id;   // 64 位：关联 chat_message.message_id
 	int _sender;
 	int _receiver;
-	int _thread_id;
+	long long _thread_id;     // 64 位
 };
 
 
@@ -65,7 +65,7 @@ public:
 	void PostTask(std::shared_ptr<FileTask> task);
 private:
 	void task_callback(std::shared_ptr<FileTask>);
-	//图片上传完成点统一处理：UpdateUploadStatus 成功后激活 pending（ZADD+EXPIRE）并尝试跨服 live 通知（计划5.7）
+	//图片上传完成点统一处理：UpdateUploadStatusWithSync 单事务（状态迁移+双方同步行）成功后尝试跨服 live 通知（计划5.7）
 	void CompleteChatImageUpload(std::shared_ptr<FileTask> task);
 	std::unordered_map<MSG_TYPES, std::function<void(std::shared_ptr<FileTask>)> > _handlers;
 	std::thread _work_thread;
